Add expression language support (#14)
Currently, the `params` used in `ObjectRoute` definitions are property
path expressions
(https://symfony.com/doc/current/components/property_access.html).

For certain use cases, more flexibility would be helpful. For example,
assume a blog post can be archived. When you link to such a blog post,
you need to include the post's year in an extra URL parameter:
`?year=...`.

Something like `#[ObjectRoute(..., params: ['year' => 'year'])]` does
not work, since the property access component cannot evaluate
conditional expressions.

This PR adds a new configuration parameter named `paramExpressions` and
uses the Symfony ExpressionLanguage component to evaluate it.

Example: `#[ObjectRoute(..., paramExpressions: ['year' =>
'this.isArchived ? this.year : null'])]`

Expressions given in `paramExpressions` can access two variables: `this`
is the object on which the route is being generated, and `params` gives
access to all parameter values. Those are the combination of the
`extraParams` passed to the object router, and all `params` evaluated
through property access expressions as previously.

In order to make optional parameters possible, the parameter name in
`paramExpressions` can be prefixed with `?` to indicate that a parameter
should not be used (filtered out) when the expression evaluates to
`null`. So, `#[ObjectRoute(..., paramExpressions: ['?year' =>
'this.isArchived ? this.year : null'])]` would only pass the `year`
parameter to the underlying router when the blog post has been archived.

Resolves #12.